feat(dynamo): assigns icon and moves nodes around
