+++ ConnectorDynamo/ConnectorDynamoFunctions/Developer/Serialization.cs
@@ -3,7 +3,7 @@ using Speckle.Core.Api;
 using Speckle.Core.Logging;
 using Speckle.Core.Models;
 
-namespace Speckle.ConnectorDynamo.Functions.Advanced
+namespace Speckle.ConnectorDynamo.Functions.Developer
 {
   public static class Serialization
   {
